+++ src/websockets/sync/connection.py
@@ -455,6 +455,7 @@ class Connection:
 
         Args:
             message: Message to send.
+            text: Force sending in a Text_ or Binary_ frame.
 
         Raises:
             ConnectionClosed: When the connection is closed.